refactor: extract contact form component
